Upgrade grafana dashboard to support recent format
When attempting to import the sample dashboard into grafana, the version is old
enough that it no longer imports properly and uses the `walk` built-in function
to create an appearance of working dashboard. This change attempts to align
most of the original queries to their newer counterparts. Summation by
`cl_name` is also used for `mongodb_metrics_document_*` metrics to simplify the
UI